+++ modules/evm/rpc/src/lib.rs
@@ -6,7 +6,11 @@ use rustc_hex::ToHex;
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_core::Bytes;
-use sp_runtime::{generic::BlockId, traits::Block as BlockT};
+use sp_runtime::{
+	codec::Codec,
+	generic::BlockId,
+	traits::{Block as BlockT, MaybeDisplay, MaybeFromStr},
+};
 
 use call_request::CallRequest;
 pub use module_evm::ExitReason;
@@ -50,12 +54,12 @@ fn error_on_execution_failure(reason: &ExitReason, data: &[u8]) -> Result<()> {
 	}
 }
 
-pub struct EVMApi<B, C> {
+pub struct EVMApi<B, C, Balance> {
 	client: Arc<C>,
-	_marker: PhantomData<B>,
+	_marker: PhantomData<(B, Balance)>,
 }
 
-impl<B, C> EVMApi<B, C> {
+impl<B, C, Balance> EVMApi<B, C, Balance> {
 	pub fn new(client: Arc<C>) -> Self {
 		Self {
 			client,
@@ -64,13 +68,14 @@ impl<B, C> EVMApi<B, C> {
 	}
 }
 
-impl<B, C> EVMApiT<B> for EVMApi<B, C>
+impl<B, C, Balance> EVMApiT<B, Balance> for EVMApi<B, C, Balance>
 where
 	B: BlockT,
 	C: ProvideRuntimeApi<B> + HeaderBackend<B> + Send + Sync + 'static,
-	C::Api: EVMRuntimeRPCApi<B>,
+	C::Api: EVMRuntimeRPCApi<B, Balance>,
+	Balance: Codec + MaybeDisplay + MaybeFromStr + Default + Send + Sync + 'static,
 {
-	fn call(&self, request: CallRequest, _: Option<B>) -> Result<Bytes> {
+	fn call(&self, request: CallRequest<Balance>, _: Option<B>) -> Result<Bytes> {
 		let hash = self.client.info().best_hash;
 
 		let CallRequest {
@@ -100,9 +105,9 @@ where
 					.map_err(|err| internal_err(format!("runtime error: {:?}", err)))?
 					.map_err(|err| internal_err(format!("execution fatal: {}", Into::<&str>::into(err))))?;
 
-				error_on_execution_failure(&info.exit_reason, &info.value)?;
+				error_on_execution_failure(&info.exit_reason, &info.output)?;
 
-				Ok(Bytes(info.value))
+				Ok(Bytes(info.output))
 			}
 			None => {
 				let info = api
@@ -116,14 +121,14 @@ where
 					.map_err(|err| internal_err(format!("runtime error: {:?}", err)))?
 					.map_err(|err| internal_err(format!("execution fatal: {}", Into::<&str>::into(err))))?;
 
-				error_on_execution_failure(&info.exit_reason, &info.value)?;
+				error_on_execution_failure(&info.exit_reason, &info.output)?;
 
-				Ok(Bytes(info.value[..].to_vec()))
+				Ok(Bytes(info.output[..].to_vec()))
 			}
 		}
 	}
 
-	fn estimate_gas(&self, request: CallRequest, _: Option<B>) -> Result<U256> {
+	fn estimate_gas(&self, request: CallRequest<Balance>, _: Option<B>) -> Result<U256> {
 		let hash = self.client.info().best_hash;
 
 		let CallRequest {
@@ -153,7 +158,7 @@ where
 					.map_err(|err| internal_err(format!("runtime error: {:?}", err)))?
 					.map_err(|err| internal_err(format!("execution fatal: {}", Into::<&str>::into(err))))?;
 
-				error_on_execution_failure(&info.exit_reason, &info.value)?;
+				error_on_execution_failure(&info.exit_reason, &info.output)?;
 
 				info.used_gas
 			}
@@ -169,7 +174,7 @@ where
 					.map_err(|err| internal_err(format!("runtime error: {:?}", err)))?
 					.map_err(|err| internal_err(format!("execution fatal: {}", Into::<&str>::into(err))))?;
 
-				error_on_execution_failure(&info.exit_reason, &info.value)?;
+				error_on_execution_failure(&info.exit_reason, &info.output)?;
 
 				info.used_gas
 			}
