+++ modules/evm/src/tests.rs
@@ -195,7 +195,7 @@ fn fail_call_return_ok() {
 			H160::default(),
 			alice(),
 			Vec::new(),
-			U256::default(),
+			0,
 			1000000,
 		));
 
@@ -204,7 +204,7 @@ fn fail_call_return_ok() {
 			H160::default(),
 			bob(),
 			Vec::new(),
-			U256::default(),
+			0,
 			1000000,
 		));
 	});
@@ -222,7 +222,7 @@ fn should_calculate_contract_address() {
 
 		let config = <Test as Trait>::config();
 
-		let handler = crate::runner::native::Handler::<Test>::new_with_precompile(
+		let handler = crate::runner::handler::Handler::<Test>::new_with_precompile(
 			&vicinity,
 			10000usize,
 			false,
@@ -266,7 +266,7 @@ fn should_create_and_call_contract() {
 		let result = <Test as Trait>::Runner::create(
 			caller.clone(),
 			contract,
-			U256::default(),
+			0,
 			1000000,
 		).unwrap();
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Returned));
@@ -285,12 +285,12 @@ fn should_create_and_call_contract() {
 			alice(),
 			contract_address,
 			multiply,
-			U256::default(),
+			0,
 			1000000
 		).unwrap();
 		assert_eq!(
 			U256::from(from_hex("0x06").unwrap().as_slice()),
-			U256::from(result.value.as_slice())
+			U256::from(result.output.as_slice())
 		);
 
 		assert_eq!(Module::<Test>::account_basic(&caller).nonce, U256::from_str("03").unwrap());
@@ -310,9 +310,9 @@ fn create_reverts_with_message() {
 	// }
 	let contract = from_hex("0x6080604052348015600f57600080fd5b5060006083576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252600d8152602001807f6572726f72206d6573736167650000000000000000000000000000000000000081525060200191505060405180910390fd5b603e8060906000396000f3fe6080604052600080fdfea265627a7a723158204741083d83bf4e3ee8099dd0b3471c81061237c2e8eccfcb513dfa4c04634b5b64736f6c63430005110032").expect("invalid hex");
 	new_test_ext().execute_with(|| {
-		let result = <Test as Trait>::Runner::create(alice(), contract, U256::default(), 12_000_000).unwrap();
+		let result = <Test as Trait>::Runner::create(alice(), contract, 0, 12_000_000).unwrap();
 		assert_eq!(result.exit_reason, ExitReason::Revert(ExitRevert::Reverted));
-		assert!(String::from_utf8_lossy(&result.value).contains("error message"));
+		assert!(String::from_utf8_lossy(&result.output).contains("error message"));
 	});
 }
 
@@ -333,7 +333,7 @@ fn call_reverts_with_message() {
 		let result = <Test as Trait>::Runner::create(
 			caller,
 			contract,
-			U256::default(),
+			0,
 			1000000,
 		).unwrap();
 
@@ -349,18 +349,18 @@ fn call_reverts_with_message() {
 			caller,
 			contract_address,
 			foo,
-			U256::default(),
+			0,
 			1000000
 		).unwrap();
 
 		assert_eq!(balance(alice()), INITIAL_BALANCE);
 		assert_eq!(result.exit_reason, ExitReason::Revert(ExitRevert::Reverted));
 		assert_eq!(
-			to_hex(&result.value, true),
+			to_hex(&result.output, true),
 			"0x8c379a00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000d6572726f72206d65737361676500000000000000000000000000000000000000"
 		);
 
-		let message  = String::from_utf8_lossy(&result.value);
+		let message  = String::from_utf8_lossy(&result.output);
 		assert!(message.contains("error message"));
 
 		assert_eq!(Module::<Test>::account_basic(&caller).nonce, U256::from_str("03").unwrap());
@@ -389,7 +389,7 @@ fn should_deploy_payable_contract() {
 			from_hex("0x000000000000000000000000000000000000000000000000000000000000007b").unwrap();
 		contract.append(&mut stored_value.clone());
 
-		let result = <Test as Trait>::Runner::create(alice(), contract, amount.into(), 100000).unwrap();
+		let result = <Test as Trait>::Runner::create(alice(), contract, amount, 100000).unwrap();
 		let contract_address = result.address;
 
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Returned));
@@ -401,13 +401,13 @@ fn should_deploy_payable_contract() {
 			alice(),
 			contract_address,
 			from_hex("0x20965255").unwrap(),
-			amount.into(),
+			amount,
 			100000,
 		)
 		.unwrap();
 
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Returned));
-		assert_eq!(result.value, stored_value);
+		assert_eq!(result.output, stored_value);
 		assert_eq!(balance(alice()), INITIAL_BALANCE - 2 * amount);
 		assert_eq!(balance(contract_address), 2 * amount);
 	});
@@ -441,16 +441,15 @@ fn should_transfer_from_contract() {
 	new_test_ext().execute_with(|| {
 		let amount = 1000u64;
 
-		let result =
-			<Test as Trait>::Runner::create(alice(), contract, 0.into(), 10000000).expect("create shouldn't fail");
+		let result = <Test as Trait>::Runner::create(alice(), contract, 0, 10000000).expect("create shouldn't fail");
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Returned));
 		let contract_address = result.address;
 
 		// send via transfer
-		let mut via_transfer = Vec::from(from_hex("0x636e082b000000000000000000000000").unwrap());
-		via_transfer.append(&mut Vec::from(H160::from(charlie()).as_bytes()));
+		let mut via_transfer = Vec::from(from_hex("0x636e082b").unwrap());
+		via_transfer.append(&mut Vec::from(H256::from(charlie()).as_bytes()));
 
-		let result = <Test as Trait>::Runner::call(alice(), contract_address, via_transfer, amount.into(), 1000000)
+		let result = <Test as Trait>::Runner::call(alice(), contract_address, via_transfer, amount, 1000000)
 			.expect("call shouldn't fail");
 
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Stopped));
@@ -458,10 +457,10 @@ fn should_transfer_from_contract() {
 		assert_eq!(balance(charlie()), 1 * amount);
 
 		// send via transfer
-		let mut via_send = from_hex("0x74be4806000000000000000000000000").unwrap();
-		via_send.append(&mut Vec::from(H160::from(charlie()).as_bytes()));
+		let mut via_send = from_hex("0x74be4806").unwrap();
+		via_send.append(&mut Vec::from(H256::from(charlie()).as_bytes()));
 
-		let result = <Test as Trait>::Runner::call(alice(), contract_address, via_send, amount.into(), 1000000)
+		let result = <Test as Trait>::Runner::call(alice(), contract_address, via_send, amount, 1000000)
 			.expect("call shouldn't fail");
 
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Stopped));
@@ -469,10 +468,10 @@ fn should_transfer_from_contract() {
 		assert_eq!(balance(alice()), INITIAL_BALANCE - 2 * amount);
 
 		// send via call
-		let mut via_call = from_hex("0x830c29ae000000000000000000000000").unwrap();
-		via_call.append(&mut Vec::from(H160::from(charlie()).as_bytes()));
+		let mut via_call = from_hex("0x830c29ae").unwrap();
+		via_call.append(&mut Vec::from(H256::from(charlie()).as_bytes()));
 
-		let result = <Test as Trait>::Runner::call(alice(), contract_address, via_call, amount.into(), 1000000)
+		let result = <Test as Trait>::Runner::call(alice(), contract_address, via_call, amount, 1000000)
 			.expect("call shouldn't fail");
 
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Stopped));
@@ -497,14 +496,12 @@ fn contract_should_deploy_contracts() {
 	// contract Contract {}
 	let contract = from_hex("0x608060405234801561001057600080fd5b5061016c806100206000396000f3fe608060405234801561001057600080fd5b506004361061002b5760003560e01c8063412a5a6d14610030575b600080fd5b61003861003a565b005b6000604051610048906100d0565b604051809103906000f080158015610064573d6000803e3d6000fd5b50905060008190806001815401808255809150509060018203906000526020600020016000909192909190916101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505050565b605b806100dd8339019056fe6080604052348015600f57600080fd5b50603e80601d6000396000f3fe6080604052600080fdfea265627a7a7231582094976cee5af14bf59c4bae67c79c12eb15de19bc18ad6038f3ee0898273c9c0564736f6c63430005110032a265627a7a72315820e19ae28dbf01eae11c526295a1ac533ea341c74d5724efe43171f6010fc98b3964736f6c63430005110032").unwrap();
 	new_test_ext().execute_with(|| {
-		let result = <Test as Trait>::Runner::create(alice(), contract.clone(), U256::default(), 1000000000).unwrap();
+		let result = <Test as Trait>::Runner::create(alice(), contract.clone(), 0, 1000000000).unwrap();
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Returned));
 
 		// Factory.createContract
 		let create_contract = from_hex("0x412a5a6d").unwrap();
-		let result =
-			<Test as Trait>::Runner::call(alice(), result.address, create_contract, U256::default(), 1000000000)
-				.unwrap();
+		let result = <Test as Trait>::Runner::call(alice(), result.address, create_contract, 0, 1000000000).unwrap();
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Stopped));
 	});
 }
@@ -528,7 +525,7 @@ fn deploy_factory() {
 	// }
 	let contract = from_hex("0x608060405234801561001057600080fd5b5060405161001d90610121565b604051809103906000f080158015610039573d6000803e3d6000fd5b506000806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055506000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663c29855786040518163ffffffff1660e01b815260040160206040518083038186803b1580156100e057600080fd5b505afa1580156100f4573d6000803e3d6000fd5b505050506040513d602081101561010a57600080fd5b81019080805190602001909291905050505061012d565b60a58061017983390190565b603e8061013b6000396000f3fe6080604052600080fdfea265627a7a7231582064177030ee644a03aaf8d65027df9e0331c8bc4b161de25bfb8aa3142848e0f864736f6c634300051100326080604052348015600f57600080fd5b5060878061001e6000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c8063c298557814602d575b600080fd5b60336049565b6040518082815260200191505060405180910390f35b6000607b90509056fea265627a7a7231582031e5a4abae00962cfe9875df1b5b0d3ce6624e220cb8c714a948794fcddb6b4f64736f6c63430005110032").unwrap();
 	new_test_ext().execute_with(|| {
-		let result = <Test as Trait>::Runner::create(alice(), contract, U256::default(), 12_000_000).unwrap();
+		let result = <Test as Trait>::Runner::create(alice(), contract, 0, 12_000_000).unwrap();
 		assert_eq!(result.exit_reason, ExitReason::Succeed(ExitSucceed::Returned));
 		assert_eq!(result.used_gas.as_u64(), 95_203u64);
 	});
