+++ runtime/acala/src/lib.rs
@@ -1627,12 +1627,12 @@ impl_runtime_apis! {
 		}
 	}
 
-	impl module_evm_rpc_runtime_api::EVMRuntimeRPCApi<Block> for Runtime {
+	impl module_evm_rpc_runtime_api::EVMRuntimeRPCApi<Block, Balance> for Runtime {
 		fn call(
 			from: H160,
 			to: H160,
 			data: Vec<u8>,
-			value: U256,
+			value: Balance,
 			gas_limit: u32,
 		) -> Result<CallInfo, sp_runtime::DispatchError> {
 			<Runtime as module_evm::Trait>::Runner::call(
@@ -1642,13 +1642,12 @@ impl_runtime_apis! {
 				value,
 				gas_limit,
 			)
-			.map_err(|err| err.into())
 		}
 
 		fn create(
 			from: H160,
 			data: Vec<u8>,
-			value: U256,
+			value: Balance,
 			gas_limit: u32,
 		) -> Result<CreateInfo, sp_runtime::DispatchError> {
 			<Runtime as module_evm::Trait>::Runner::create(
@@ -1657,7 +1656,6 @@ impl_runtime_apis! {
 				value,
 				gas_limit,
 			)
-			.map_err(|err| err.into())
 		}
 
 	}
